requirements.txt
langchain==0.0.352
langchain-openai
langchain-community
openai
chromadb
pyyaml
bs4
tiktoken
python-dotenv